Fix PS1 analog directions: compose stick values from per-direction bindings
Trusting the raw SDL axis sign at play time folded up/left into down/right
on the 8BitDo Pro 2 (BT). Analog values are now composed from the Controls
panel's per-direction bindings (plus-half minus minus-half), the same model
RetroArch and DuckStation use — direction comes from the captured binding,
never the runtime axis sign. Direct-axis read remains the fallback for
consoles with no saved analog bindings.

Also: input-state callback returns sign-extended int (interop hygiene for
clang-compiled cores that read the full return register).

+++ src/Emutastic/Emulator/EmulatorSession.cs
@@ -1336,7 +1336,7 @@ namespace Emutastic.Emulator
         // ── In-game disc switching (L3 + Start chord) ───────────────────────────────────────────────
         // Wraps SdlInput.GetInputState so we can inject a JOYPAD_L press on port 0 for FDS "disk side
         // change" (FDS cores don't expose the disk-control interface — they read an L press instead).
-        private short InputState_cb(uint port, uint device, uint index, uint id)
+        private int InputState_cb(uint port, uint device, uint index, uint id)
         {
             if (port == 0 && device == SdlInput.RETRO_DEVICE_JOYPAD
                 && id == LibretroInput.JOYPAD_L && _fdsSideChangeFrames > 0)
